int64 to int

+++ properloading/scrollPageDash.go
@@ -14,7 +14,7 @@ import (
 
 //ScrollPageDash ...
 //Scrolls through docs being loaded
-func ScrollPageDash(eclient *elastic.Client, docIDs []string, scrollID string) (string, []types.JournalEntry, int64, error) {
+func ScrollPageDash(eclient *elastic.Client, docIDs []string, scrollID string) (string, []types.JournalEntry, int, error) {
 
 	ctx := context.Background()
 
@@ -54,16 +54,16 @@ func ScrollPageDash(eclient *elastic.Client, docIDs []string, scrollID string) (
 		head, err := uses.ConvertEntryToJournalEntry(eclient, hit.Id, false)
 		arrResults = append(arrResults, head)
 		if err != nil {
-			return res.ScrollId, arrResults, res.Hits.TotalHits, errors.New("ISSUE WITH CONVERT FUNCTION")
+			return res.ScrollId, arrResults, int(res.Hits.TotalHits), errors.New("ISSUE WITH CONVERT FUNCTION")
 
 		}
 
 		if err == io.EOF {
-			return res.ScrollId, arrResults, res.Hits.TotalHits, errors.New("Out of bounds")
+			return res.ScrollId, arrResults, int(res.Hits.TotalHits), errors.New("Out of bounds")
 
 		}
 
 	}
 
-	return res.ScrollId, arrResults, res.Hits.TotalHits, err
+	return res.ScrollId, arrResults, int(res.Hits.TotalHits), err
 }
